windows-builder: copy via GCE bucket
This commit adds logic to try and use a bucket to transfer the workspace
to the Windows VM. It creates an object in the {project-id}_cloudbuild
bucket and then instructs the Windows VM to use `gsutil` via Powershell
to download the given object.

This assumes the Windows VM has read permission for the bucket.

If this process fails, it fallsback to the existing mechanism of
transferring files.

fixes #375